requirements.txt
pydantic
PyPDF2
google-generativeai
streamlit
python-dotenv
